callbacks separated into multiple files to accommodate optional installation of Plots
- Readme updated : included instructions for installation of Plots
- modified option plots to setup file
- implemented try-catch to verify if Plots is installed
- imports implicated in previous changes updated.
