+++ .travis.yml
@@ -8,7 +8,7 @@ python:
   - "3.6"
   - "3.7"
 install:
-  - pip install -e .[all]
+  - pip install -e .[all,plots]
   - pip install pytest-cov
 script:
   - pytest --cov=mushroom_rl --cov-report=xml
