+++ mushroom_rl/utils/plots/__init__.py
@@ -1,7 +1,22 @@
-from .plot_item_w_buffer import PlotItemWBuffer
-from .databuffer import DataBuffer
-from .window import Window
-from ._implementations import common_plots
-from ._implementations import common_buffers
+__extras__ = []
 
-__all__ = ['DataBuffer', 'PlotItemWBuffer', 'Window', 'common_plots', 'common_buffers']
+try:
+    from .plot_item_w_buffer import PlotItemWBuffer
+    __extras__.append('PlotItemWBuffer')
+
+    from .databuffer import DataBuffer
+    __extras__.append('DataBuffer')
+
+    from .window import Window
+    __extras__.append('Window')
+
+    from ._implementations import common_plots
+    __extras__.append('common_plots')
+
+    from ._implementations import common_buffers
+    __extras__.append('common_buffers')
+
+except ImportError:
+    pass
+
+__all__ = __extras__
