+++ tests/utils/test_callbacks.py
@@ -4,6 +4,7 @@ from mushroom_rl.algorithms.value import SARSA
 from mushroom_rl.policy import EpsGreedy
 from mushroom_rl.utils.parameters import Parameter, ExponentialParameter
 from mushroom_rl.utils.callbacks import *
+import numpy as np
 
 
 def test_collect_dataset():
